Split CoreGinModule to several GIN modules
