+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/clipboard/ZeroClipboardInjector.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.ide.core;
+package org.eclipse.che.ide.clipboard;
 
 import org.eclipse.che.ide.api.component.Component;
 import org.eclipse.che.ide.util.loging.Log;
