+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/factory/FactoryServiceClientImpl.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.ide.api.factory;
+package org.eclipse.che.ide.factory;
 
 import com.google.common.base.Joiner;
 import com.google.gwt.http.client.RequestBuilder;
@@ -19,6 +19,7 @@ import org.eclipse.che.api.factory.shared.dto.FactoryDto;
 import org.eclipse.che.api.promises.client.Promise;
 import org.eclipse.che.commons.annotation.Nullable;
 import org.eclipse.che.ide.MimeType;
+import org.eclipse.che.ide.api.factory.FactoryServiceClient;
 import org.eclipse.che.ide.rest.AsyncRequestCallback;
 import org.eclipse.che.ide.rest.AsyncRequestFactory;
 import org.eclipse.che.ide.rest.DtoUnmarshallerFactory;
