+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/oauth/OAuth2AuthenticatorRegistryImpl.java
@@ -13,8 +13,8 @@ package org.eclipse.che.ide.oauth;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 
-import org.eclipse.che.ide.api.oauth.OAuth2AuthenticatorRegistry;
 import org.eclipse.che.ide.api.oauth.OAuth2Authenticator;
+import org.eclipse.che.ide.api.oauth.OAuth2AuthenticatorRegistry;
 import org.eclipse.che.ide.util.loging.Log;
 
 import java.util.HashMap;
