+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/user/UserProfileServiceClientImpl.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.ide.api.user;
+package org.eclipse.che.ide.user;
 
 import com.google.inject.Inject;
